change BooleanSettingsFieldViewModel folder and namespace
